+++ webpack.conf.js
@@ -1,20 +1,8 @@
-/* global module:false, __dirname:false */
-
 var webpack = require('webpack');
 
 module.exports = {
-    devtool: 'cheap-module-source-map',
-    module: {
-        loaders: [
-            {
-                test:    /\.js$/,
-                exclude: /node_modules/,
-                loaders: ['babel-loader']
-            }
-        ]
-    },
 
     resolve: {
-        extensions: ['.js', '.jsx', '.json']
+        extensions: ['.js']
     }
 };